qute_core/qute_core/demo/qute_xxx_test/requirements.txt
configparser==3.7.4
PyMySQL==0.7.11
requests==2.21.0
pytest==5.0.1
pytest-xdist==1.29.0
allure-python-commons==2.7.1
pytest-allure-adaptor2==1.7.11
pytest-html==1.21.1
redis==2.10.6
pymongo==3.8.0
oss2==2.7.0
elasticsearch==7.0.2